pyproject.toml
[project]
name = "movie-rag-system"
version = "1.0.0"
description = "AI-powered movie recommendation system using RAG (Retrieval-Augmented Generation)"
authors = [
    {name = "Raghav Goyal", email = "gyl.rghv@gmail.com"}
]

keywords = ["rag", "movie", "recommendation", "ai", "postgresql", "chromadb"]

readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.32.3",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "chromadb>=0.4.0",
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "sqlalchemy>=2.0.43",
    "python-dateutil>=2.9.0.post0",
    "psycopg2-binary>=2.9.10",
    "openai>=1.107.3",
    "httpx>=0.28.1",
    "langchain-community>=0.3.29",
    "langchain-chroma>=0.2.6",
    "plotly>=6.3.0",
]

[project.scripts]
movie-rag = "main:main"
movie-rag-web = "run_streamlit:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0", 
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/movie_rag"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
